httphdr: moved to option.go
