+++ httphdr/httphdr.go
@@ -3,8 +3,6 @@ package httphdr
 import (
 	"bytes"
 	"net/http"
-	"regexp"
-	"strings"
 )
 
 func PutHTTPHeaders(dst http.Header, src http.Header, keys ...string) {
@@ -25,44 +23,3 @@ func SizeOfHTTPHeader(header http.Header) int {
 	_ = header.Write(buf)
 	return buf.Len()
 }
-
-type Option struct {
-	Name       string
-	Parameters map[string]string
-}
-
-func ParseOptions(directive string) []*Option {
-	result := make([]*Option, 0, 128)
-	var option *Option
-	matches := optionRgx.FindAllString(directive, -1)
-	for _, match := range matches {
-		if strings.HasSuffix(match, ";") {
-			if option != nil {
-				result = append(result, option)
-			}
-			option = &Option{
-				Name:       strings.ToLower(strings.TrimSpace(strings.TrimSuffix(match, ";"))),
-				Parameters: make(map[string]string),
-			}
-			continue
-		}
-		if option == nil {
-			option = &Option{
-				Name:       "",
-				Parameters: make(map[string]string),
-			}
-		}
-		var key, value string
-		key = match
-		if index := strings.Index(match, "="); index != -1 {
-			key, value = match[:index], match[index+1:]
-		}
-		option.Parameters[strings.ToLower(key)] = strings.TrimSpace(value)
-	}
-	if option != nil {
-		result = append(result, option)
-	}
-	return result
-}
-
-var optionRgx = regexp.MustCompile(`([a-zA-Z][a-zA-Z_\-]*(?:\s*;)?)(?:=(?:"([^"]*)"|([^ \t",;]*)))?`)
